I solved it but it.
BUT

It only works for small use cases. It will fail with large arrays. I have to find a way to reduce the runtime or it will fail.

The MDN says that the reduce method doesn't have this problem but that using Math.max(...arr) does because of how the function works